+dev-massigy-threads-smaphores
Semaphores works for some mazes. In other mazes, we have two problems,

first, is sometimes that semaphore initial value is not enough, so basicaly there is not enough threads
which causes a deadlock.

second, sometimes with some mazes there is a seg fault which I do not know where it does come from.