+dev-massigy-threads-semaphores: threads limit mechanism overall done.
In these last commit here is the digest:

First, I've decided to use semaphores since they can be initialized to a value and the need
here is all about waiting for a thread to finish if all the threads pool is consumed, and as soon
as one thread finishes, we post/signal it to the thread pool.

In other terms, I've decided to use semaphores rather then mutexs since we are not trying to protect ownership over
a shared ressource but we are only defining our threads pool length.

Also, I've tried to count how much the `solveMaze_threaded` uses in terms of threads, and this results are queit useless since
we are not defining a boundry on how much threads the function should use by the end of its execution, but we are only defining a boundry
of how much threads the function can use at the same time! ( which are completely diffrent mesures )

Finaly, I've come to the conclusion that the system is working, since when decrementing too much the boundry, the program execution
basically freezes and we enter a deadlock situation.

@Massiles.