+dev-massigy-threads: removed all mutex related implementations.
I've decided to go with the semaphores instead of mutexes, since we only need to fix a limit on how much threads the function
`solveMaze_threaded` can use at the same time.

So, since there is no master ressources envolved and no ownership envolved also, mutexes are queit useless in this situation, and also
not forgetting the fact that mutexes can only be manipulated by the same thread.

for more infos checkout my commit on the dev-massigy-threads-semaphores done around the same hour.

@Massiles.